+++ indispensable/src/commonMain/kotlin/at/asitplus/signum/indispensable/X509SignatureAlgorithm.kt
@@ -96,8 +96,13 @@ open class X509SignatureAlgorithm(
             else -> throw IllegalArgumentException("Unsupported hash algorithm.")
         }
 
+    /**
+     * **Accessing this property may throw.**
+     * @throws IllegalArgumentException in case a non-supported [X509SignatureAlgorithm] is to be converted
+     */
     // TODO update when core signature data classes become extensible
     override val algorithm: SignatureAlgorithm
+        @Throws(IllegalArgumentException::class)
         get() = when (this) {
             is EC -> SignatureAlgorithm.ECDSA(digest, null)
             is RSA -> SignatureAlgorithm.RSA(digest, if (pssBits != null) RSAPadding.PSS else RSAPadding.PKCS1)